app/requirements.txt
ultralytics==8.1.5
matplotlib==3.8.2
numpy==1.24.3 
streamlit==1.31.0
pillow==10.0.1 
opencv-python==4.9.0.80




